requirements.txt
matplotlib==3.8.4
numpy==1.26.4
opencv_python==4.9.0.80
Pillow==10.3.0
segmentation_models_pytorch==0.3.3
streamlit==1.34.0
streamlit_drawable_canvas==0.9.3
torch==2.3.1
torchvision==0.18.1
